refactor: Injector에서 공통 로직을 수행하도록 변경
